visualization/requirements.txt
pandas
psycopg2-binary
matplotlib
seaborn
